fix(fiber): Fix fiber returning 500 internal server error
https://github.com/awslabs/aws-lambda-go-api-proxy/commit/fde59295b7aa31f7d8f65997da1a94bad8f69303 caused a regression
in the fiber adapter logic as we previously would try to get the source IP from nil, which
internally would just return an empty string.

Since above commit, the RemoteAddr is not nil anymore, but is lacking the required port
that is needed in order to parse it as a *TCPAddr

This commit makes sure the port section is always added if it does not exist. Furthermore, it
adds a test that prevents this from happening again and we now have logging in place.